kansai-housing-intel/Crossactor/ceo_system/backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
httpx>=0.27.0
line-bot-sdk>=3.0.0
python-multipart>=0.0.9
